Feature user profile

+++ src/shared/container/providers/StorageProvider/implementations/LocalStorageProvider.ts
@@ -2,6 +2,7 @@ import fs from "fs";
 import { resolve } from "path";
 
 import upload from "@config/upload";
+import { ApplicationError } from "@shared/errors/ApplicationError";
 
 import { IStorageProvider } from "../IStorageProvider";
 
@@ -18,7 +19,10 @@ class LocalStorageProvider implements IStorageProvider {
   async delete(file: string, folder: string): Promise<void> {
     const filename = resolve(`${upload.tmpFolder}/${folder}`, file);
 
-    const fileExists = await fs.promises.stat(filename);
+    const fileExists = fs.existsSync(filename);
+
+    if (!fileExists)
+      throw new ApplicationError("File or directory not found", 404);
 
     if (fileExists) await fs.promises.unlink(filename);
   }
